requirements.txt
anyio==4.9.0
asgiref==3.8.1
autopep8==2.3.2
black==25.1.0
cfgv==3.4.0
click==8.1.8
coverage==7.8.0
distlib==0.3.9
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
factory_boy==3.3.3
Faker==37.3.0
filelock==3.18.0
flake8==7.2.0
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.16.0
identify==2.6.10
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
libcst==1.7.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
promise==2.3
psycopg==3.2.9
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-django==4.11.1
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
rich==14.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
strawberry-graphql==0.270.0
strawberry-graphql-django==0.59.1
text-unidecode==1.3
typer==0.15.4
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
virtualenv==20.31.2
xlwt==1.3.0
